requirements.txt
altonomy-client==1.0.3
altonomy-services>=2.0.8
zerorpc
pandas
numpy
sqlalchemy==1.4.41
requests
mysql-connector==2.2.9
altonomy-loggers>=0.1.2
altonomy-models==0.6.22
flake8
pytest
pytest-cov
pytest-mock
fakeredis
sphinx
altonomy-exchanges==1.0.150
cachetools
altonomy-ref-data-api==0.0.24
nested-lookup
altonomy-price-server-api>=0.0.5
pyzmq==22.3.0
